Move lists under management list
